+++ buildSrc/src/main/java/Config.kt
@@ -33,7 +33,7 @@ object Config {
     object Libs {
         val kotlinStd = "org.jetbrains.kotlin:kotlin-stdlib"
 
-        val sentryJavaVersion = "8.14.0"
+        val sentryJavaVersion = "8.15.1"
         val sentryAndroid = "io.sentry:sentry-android:$sentryJavaVersion"
         val sentryJava = "io.sentry:sentry:$sentryJavaVersion"