+++ CHANGELOG.md
@@ -8,7 +8,7 @@
 
 ### Fixes
 
-- `beforeSend` overriding default values even if they were not set explicitly ([#376](https://github.com/getsentry/sentry-kotlin-multiplatform/pull/376))
+- `beforeSend` overriding default release even if it was not set explicitly ([#376](https://github.com/getsentry/sentry-kotlin-multiplatform/pull/376))
 
 ### Dependencies