+++ CHANGELOG.md
@@ -6,6 +6,12 @@
 
 - Gradle Plugin: implement conditional Cocoa linking for targets ([#421](https://github.com/getsentry/sentry-kotlin-multiplatform/pull/421), [#423](https://github.com/getsentry/sentry-kotlin-multiplatform/pull/423))
 
+### Dependencies
+
+- Bump Cocoa SDK from v8.53.1 to v8.53.2 ([#419](https://github.com/getsentry/sentry-kotlin-multiplatform/pull/419))
+  - [changelog](https://github.com/getsentry/sentry-cocoa/blob/main/CHANGELOG.md#8532)
+  - [diff](https://github.com/getsentry/sentry-cocoa/compare/8.53.1...8.53.2)
+
 ## 0.14.0
 
 ### Dependencies